Restore Nuget packages on yarn build

+++ vnext/Scripts/NuGetRestoreForceEvaluateAllSolutions.ps1
@@ -1,4 +1,5 @@
 param(
+    [switch] $SkipLockDeletion
 )
 
 [string] $RepoRoot = Resolve-Path "$PSScriptRoot\..\.."
@@ -6,16 +7,17 @@ param(
 $StartingLocation = Get-Location
 Set-Location -Path $RepoRoot
 
-try
-{
-    # Delete existing lock files
-    $existingLockFiles = (Get-ChildItem -File -Recurse -Path $RepoRoot -Filter *.lock.json)
-    $existingLockFiles | Foreach-Object {
-        Write-Host Deleting $_.FullName
-        Remove-Item $_.FullName
+try {
+    if (-not $SkipLockDeletion) {
+        # Delete existing lock files
+        $existingLockFiles = (Get-ChildItem -File -Recurse -Path $RepoRoot -Filter *.lock.json)
+        $existingLockFiles | Foreach-Object {
+            Write-Host Deleting $_.FullName
+            Remove-Item $_.FullName
+        }
     }
 
-    $packagesSolutions = (Get-ChildItem -File -Recurse -Path $RepoRoot\packages -Filter *.sln )| Where-Object { !$_.FullName.Contains('node_modules') -and !$_.FullName.Contains('e2etest') }
+    $packagesSolutions = (Get-ChildItem -File -Recurse -Path $RepoRoot\packages -Filter *.sln ) | Where-Object { !$_.FullName.Contains('node_modules') -and !$_.FullName.Contains('e2etest') }
     $vnextSolutions = (Get-ChildItem -File -Path $RepoRoot\vnext -Filter *.sln)
 
     # Run all solutions with their defaults
@@ -31,7 +33,6 @@ try
         & msbuild /t:Restore /p:RestoreForceEvaluate=true /p:UseExperimentalWinUI3=true $_.FullName
     }
 }
-finally
-{
+finally {
     Set-Location -Path "$StartingLocation"
 }
